requirements.txt
djangorestframework-simplejwt
django-cors-headers
django
djangorestframework
django-filter
djangorestframework-recursive
psycopg2
django-mptt
faker
